+++ python-connectors/sharepoint-online_lists/connector.py
@@ -2,8 +2,8 @@ from dataiku.connector import Connector
 
 from sharepoint_client import SharePointClient
 from sharepoint_constants import SharePointConstants
-from sharepoint_lists import assert_list_title, get_dss_type
-from sharepoint_lists import SharePointListWriter, column_ids_to_names, sharepoint_to_dss_date
+from sharepoint_lists import assert_list_title
+from sharepoint_lists import column_ids_to_names, sharepoint_to_dss_date
 from common import parse_query_string_to_dict
 from safe_logger import SafeLogger
 from dss_constants import DSSConstants
@@ -16,16 +16,10 @@ class SharePointListsConnector(Connector):
 
     def __init__(self, config, plugin_config):
         Connector.__init__(self, config, plugin_config)
-        logger.info('SharePoint Online plugin connector v1.0.14')
+        logger.info('SharePoint Online plugin connector v{}'.format(DSSConstants.PLUGIN_VERSION))
         self.sharepoint_list_title = self.config.get("sharepoint_list_title")
         self.auth_type = config.get('auth_type')
         logger.info('init:sharepoint_list_title={}, auth_type={}'.format(self.sharepoint_list_title, self.auth_type))
-        self.column_ids = {}
-        self.column_names = {}
-        self.column_entity_property_name = {}
-        self.columns_to_format = []
-        self.dss_column_name = {}
-        self.column_sharepoint_type = {}
         self.expand_lookup = config.get("expand_lookup", False)
         self.metadata_to_retrieve = config.get("metadata_to_retrieve", [])
         advanced_parameters = config.get("advanced_parameters", False)
@@ -44,64 +38,16 @@ class SharePointListsConnector(Connector):
         self.client = SharePointClient(config)
         self.sharepoint_list_view_id = None
         if self.sharepoint_list_view_title:
-            self.sharepoint_list_view_id = self.get_view_id(self.sharepoint_list_title, self.sharepoint_list_view_title)
-
-    def get_view_id(self, list_title, view_title):
-        if not list_title:
-            return None
-        views = self.client.get_list_views(list_title)
-        for view in views:
-            if view.get("Title") == view_title:
-                return view.get("Id")
-        raise ValueError("View '{}' does not exist in list '{}'.".format(view_title, list_title))
+            self.sharepoint_list_view_id = self.client.get_view_id(self.sharepoint_list_title, self.sharepoint_list_view_title)
 
     def get_read_schema(self):
-        logger.info('get_read_schema')
-        sharepoint_columns = self.client.get_list_fields(self.sharepoint_list_title)
-        dss_columns = []
-        self.column_ids = {}
-        self.column_names = {}
-        self.column_entity_property_name = {}
-        self.columns_to_format = []
-        for column in sharepoint_columns:
-            logger.info("get_read_schema:{}/{}/{}/{}/{}/{}".format(
-                column[SharePointConstants.TITLE_COLUMN],
-                column[SharePointConstants.TYPE_AS_STRING],
-                column[SharePointConstants.STATIC_NAME],
-                column[SharePointConstants.INTERNAL_NAME],
-                column[SharePointConstants.ENTITY_PROPERTY_NAME],
-                self.is_column_displayable(column)
-            ))
-            if self.is_column_displayable(column):
-                sharepoint_type = get_dss_type(column[SharePointConstants.TYPE_AS_STRING])
-                self.column_sharepoint_type[column[SharePointConstants.STATIC_NAME]] = column[SharePointConstants.TYPE_AS_STRING]
-                if sharepoint_type is not None:
-                    dss_columns.append({
-                        SharePointConstants.NAME_COLUMN: column[SharePointConstants.TITLE_COLUMN],
-                        SharePointConstants.TYPE_COLUMN: sharepoint_type
-                    })
-                    self.column_ids[column[SharePointConstants.STATIC_NAME]] = sharepoint_type
-                    self.column_names[column[SharePointConstants.STATIC_NAME]] = column[SharePointConstants.TITLE_COLUMN]
-                    self.column_entity_property_name[column[SharePointConstants.STATIC_NAME]] = column[SharePointConstants.ENTITY_PROPERTY_NAME]
-                    self.dss_column_name[column[SharePointConstants.STATIC_NAME]] = column[SharePointConstants.TITLE_COLUMN]
-                    self.dss_column_name[column[SharePointConstants.ENTITY_PROPERTY_NAME]] = column[SharePointConstants.TITLE_COLUMN]
-                if sharepoint_type == "date":
-                    self.columns_to_format.append((column[SharePointConstants.STATIC_NAME], sharepoint_type))
-        logger.info("get_read_schema: Schema updated with {}".format(dss_columns))
-        return {
-            SharePointConstants.COLUMNS: dss_columns
-        }
+        return self.client.get_read_schema(display_metadata=self.display_metadata, metadata_to_retrieve=self.metadata_to_retrieve)
 
     @staticmethod
     def get_column_lookup_field(column_static_name):
         if column_static_name in SharePointConstants.EXPENDABLES_FIELDS:
             return SharePointConstants.EXPENDABLES_FIELDS.get(column_static_name)
         return None
-
-    def is_column_displayable(self, column):
-        if self.display_metadata and (column['StaticName'] in self.metadata_to_retrieve):
-            return True
-        return (not column[SharePointConstants.HIDDEN_COLUMN])
 
     @staticmethod
     def must_column_display_be_forced(column):
@@ -114,8 +60,8 @@ class SharePointListsConnector(Connector):
 
     def generate_rows(self, dataset_schema=None, dataset_partitioning=None,
                       partition_id=None, records_limit=-1):
-        if self.column_ids == {}:
-            self.get_read_schema()
+        if self.client.column_ids == {}:
+            self.client.get_read_schema()
 
         logger.info('generate_row:dataset_schema={}, dataset_partitioning={}, partition_id={}, records_limit={}'.format(
             dataset_schema, dataset_partitioning, partition_id, records_limit
@@ -134,7 +80,7 @@ class SharePointListsConnector(Connector):
             rows = self.get_page_rows(page)
             for row in rows:
                 row = self.format_row(row)
-                yield column_ids_to_names(self.dss_column_name, row)
+                yield column_ids_to_names(self.client.dss_column_name, row)
             record_count += len(rows)
             if is_record_limit and record_count >= records_limit:
                 break
@@ -159,7 +105,7 @@ class SharePointListsConnector(Connector):
         return page.get("Row", "")
 
     def format_row(self, row):
-        for column_to_format, type_to_process in self.columns_to_format:
+        for column_to_format, type_to_process in self.client.columns_to_format:
             value = row.get(column_to_format)
             if value:
                 row[column_to_format] = sharepoint_to_dss_date(value)
@@ -168,16 +114,7 @@ class SharePointListsConnector(Connector):
     def get_writer(self, dataset_schema=None, dataset_partitioning=None,
                    partition_id=None):
         assert_list_title(self.sharepoint_list_title)
-        return SharePointListWriter(
-            self.config,
-            self,
-            dataset_schema,
-            dataset_partitioning,
-            partition_id,
-            max_workers=self.max_workers,
-            batch_size=self.batch_size,
-            write_mode=self.write_mode
-        )
+        return self.client.get_writer(dataset_schema, dataset_partitioning, partition_id, self.max_workers, self.batch_size, self.write_mode)
 
     def get_partitioning(self):
         logger.info('get_partitioning')