+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [Version 1.1.0](https://github.com/dataiku/dss-plugin-sharepoint-online/releases/tag/v1.1.0) - Feature release - 2023-11-10
+
+- Adding an **Append to list** recipe
+- Updated code-env descriptor for DSS 12
+
+## [Version 1.0.15](https://github.com/dataiku/dss-plugin-sharepoint-online/releases/tag/v1.0.15) - Feature release - 2023-11-10
+
+- Adding an **Append to list** recipe
+
 ## [Version 1.0.14](https://github.com/dataiku/dss-plugin-sharepoint-online/releases/tag/v1.0.14) - Bugfix release - 2023-04-18
 
 - Updated code-env descriptor for DSS 12